requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
Django==1.9.5
djangorestframework==3.3.3
gunicorn==19.4.5
meld3==1.0.2
singledispatch==3.4.0.3
six==1.10.0
supervisor==3.2.3
tornado==4.3
virtualenv==15.0.1
configobj==5.0.6